pyproject.toml
[project]
name = "mri-phen"
version = "0.1.0"
description = "Project for MRI Phenotyping using Deep Learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "huggingface-hub>=1.2.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydicom>=3.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
    "wheel>=0.45.1",
    "xlrd>=2.0.2",
]

[tool.uv]
package = false
